Tweak heuristic properties.

+++ board.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 
 class Board:
     def __init__(self):
-        self.board = np.zeros((8,8))
+        self.board = np.zeros((8,8), dtype=np.int32)
         self.board[3][4] = BLK
         self.board[4][3] = BLK
         self.board[3][3] = WHT
